feat(207-01): narrow day-index types to uint32, convert ticketWriteSlot to bool
- purchaseStartDay, dailyIdx, lastDailyJackpotDay: uint48 -> uint32
- All day-index mapping keys (12 mappings): uint48 -> uint32
- ticketWriteSlot: uint8 -> bool with negation toggle
- _tqWriteKey/_tqReadKey: direct bool reads (no != 0 / == 0)
- GameTimeLib: returns uint32, JACKPOT_RESET_TIME uint32
- _simulatedDayIndex/_simulatedDayIndexAt: return uint32
- _currentMintDay/_isDistressMode: use uint32 locals

+++ contracts/libraries/GameTimeLib.sol
@@ -11,14 +11,14 @@ import {ContractAddresses} from "../ContractAddresses.sol";
  */
 library GameTimeLib {
     /// @notice Daily reset time in seconds from midnight UTC (22:57 UTC = 82620 seconds)
-    uint48 internal constant JACKPOT_RESET_TIME = 82620;
+    uint32 internal constant JACKPOT_RESET_TIME = 82620;
 
     /**
      * @notice Get current day index relative to deploy time.
      * @dev Day 1 = deploy day. Days reset at JACKPOT_RESET_TIME (22:57 UTC).
      * @return Current day index (1-indexed from deploy day).
      */
-    function currentDayIndex() internal view returns (uint48) {
+    function currentDayIndex() internal view returns (uint32) {
         return currentDayIndexAt(uint48(block.timestamp));
     }
 
@@ -28,8 +28,8 @@ library GameTimeLib {
      * @param ts Timestamp to evaluate.
      * @return Day index (1-indexed from deploy day).
      */
-    function currentDayIndexAt(uint48 ts) internal pure returns (uint48) {
-        uint48 currentDayBoundary = uint48((ts - JACKPOT_RESET_TIME) / 1 days);
+    function currentDayIndexAt(uint48 ts) internal pure returns (uint32) {
+        uint32 currentDayBoundary = uint32((ts - JACKPOT_RESET_TIME) / 1 days);
         return currentDayBoundary - ContractAddresses.DEPLOY_DAY_BOUNDARY + 1;
     }
 }
